v1.1.2 正式版发布（release）

+++ xingtools-core/src/main/java/top/cutexingluo/tools/common/valid/num/longstatus/LongRangeData.java
@@ -22,13 +22,13 @@ public class LongRangeData {
      *
      * <p> min value (inclusive)</p>
      */
-    protected double min;
+    protected long min;
     /**
      * 区间最大值
      *
      * <p>max value (inclusive)</p>
      */
-    protected double max;
+    protected long max;
 
     /**
      * 转化注解
